# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Logueo",
        "description": "Endpoints para gestión de logueo",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/partidos": {
            "get": {
                "tags": [
                    "Partidos"
                ],
                "summary": "Obtener todos los partidos",
                "description": "Obtener todos los centros.",
                "operationId": "b43fc01333c208a642fc831977f258bc",
                "responses": {
                    "200": {
                        "description": "Lista de partidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "partidos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Partido"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Partidos"
                ],
                "summary": "Crear un nuevo partido",
                "description": "Crear un nuevo centro.",
                "operationId": "0f5ffa8bee14eff46ddd306771264556",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Partido"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partido creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partido creado correctamente"
                                        },
                                        "partido": {
                                            "$ref": "#/components/schemas/Partido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    }
                }
            }
        },
        "/api/partidos/{partido}": {
            "get": {
                "tags": [
                    "Partidos"
                ],
                "summary": "Obtener un partido por su ID",
                "description": "Obtener un centro por su ID.",
                "operationId": "0e0ff8595f76aee308b97f71fe05cd1e",
                "parameters": [
                    {
                        "name": "partido",
                        "in": "path",
                        "description": "ID del partido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partido encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "partido": {
                                            "$ref": "#/components/schemas/Partido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partido no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Partidos"
                ],
                "summary": "Actualizar un partido existente",
                "description": "Actualizar un partido existente.",
                "operationId": "63ca8d98a4f70dd30bc532b97f0e6239",
                "parameters": [
                    {
                        "name": "partido",
                        "in": "path",
                        "description": "ID del partido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Partido"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partido actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partido actualizado correctamente"
                                        },
                                        "partido": {
                                            "$ref": "#/components/schemas/Partido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    },
                    "404": {
                        "description": "Partido no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Partidos"
                ],
                "summary": "Eliminar un partido",
                "description": "Eliminar un partido.",
                "operationId": "6ca749b06d0ef378d9451bc64bf04895",
                "parameters": [
                    {
                        "name": "partido",
                        "in": "path",
                        "description": "ID del partido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partido eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partido eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partido no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Partido": {
                "title": "Partido",
                "properties": {
                    "id": {
                        "description": "ID único del ciclo",
                        "type": "integer",
                        "example": 1
                    },
                    "equipo_local_id": {
                        "description": "ID del equipo local",
                        "type": "integer",
                        "example": 1
                    },
                    "equipo_visitante_id": {
                        "description": "ID del equipo visitante",
                        "type": "integer",
                        "example": 2
                    },
                    "fecha": {
                        "description": "Fecha del partido",
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-15"
                    },
                    "tiempo": {
                        "description": "Minuto del partido",
                        "type": "integer",
                        "example": 16
                    },
                    "goles_local": {
                        "description": "Goles del equipo local",
                        "type": "integer",
                        "example": 3
                    },
                    "goles_visitante": {
                        "description": "Goles del equipo visitante",
                        "type": "integer",
                        "example": 2
                    },
                    "pabellon_id": {
                        "description": "ID del pabellón donde se juega el partido",
                        "type": "integer",
                        "example": 101
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}